requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
PyPDF2==3.0.1
python-docx==1.1.2
typing_extensions==4.12.2
Werkzeug==3.0.4
